pyproject.toml
[tool.poetry]
name = "drofji-snapadmin"
version = "0.1.0a1"
description = "Automatic and customizable admin interface for Django projects."
authors = ["Alexander Wiese"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/drofji/django-snapadmin"
repository = "https://github.com/drofji/django-snapadmin"
keywords = ["django", "admin", "ui", "Snap", "extra-settings"]

packages = [{include = "snapadmin"}]

include = [
    "snapadmin/templates/**/*",
    "snapadmin/static/**/*",
]

[tool.poetry.dependencies]
python = ">=3.10"
Django = ">=5.2"
python-dotenv = ">=1.0.0"
django-admin-rangefilter = ">=0.13.3"
django-unfold = ">=0.40.0"
django-ckeditor-5 = ">=0.0.12"
django-colorfield = ">=0.11.0"
django-cors-headers = ">=4.4.0"
django-admin-autocomplete-filter = ">=0.7.1"
django-extra-settings = ">=0.12.0"
djangorestframework = ">=3.15.0"
drf-spectacular = ">=0.27.0"
djangorestframework-simplejwt = ">=5.3.0"
structlog = ">=24.1.0"
colorama = ">=0.4.6"
graphene-django = ">=3.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
flake8 = "^7.0.0"
pytest = "^8.0.0"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
